Launcher3: Move restart check to home page
It's just a flag check. And avoids having to implement a listener for the home key (and still miss home gesture).
Also, restarting the launcher when navigating to previous settings page doesn't seem right.
This way, the launcher is restarted once, if needed, when user goes back to home page, which is where the changes are visible, anyway.

Change-Id: Ie7f0a8d015d0e47bfb93a4920fa591e237f43d40
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Mohammad Hasan Keramat J <ikeramat@protonmail.com>